#78 [refactor] : 외부 API 호출은 도메인 내에서 interface, 인프라에서 구현하는 방식으로 변경한다
